+++ src/lib/auth.tsx
@@ -1,4 +1,4 @@
-import { AuthForm, AuthForm2 } from "../../interfaces/auth"
+import { AuthForm, AuthForm2 } from "../interfaces/auth"
 
 export const formBody = (
 { form, form2, authPage }: { form: AuthForm; form2: AuthForm2; authPage: string }): Record<string, any> => {
